move all bluetooth init code into main.c
